Module squid_vm::instructiondefs

Defines all the instructions.

Constants

AVP

Add variable pointer.

D_VFD

Debug variable from data.

D_VFS

Debug variable from stack.

F_ADD

Adds two floats from the stack.

F_DVD

Divides two floats from the stack.

F_EXP

Take two floats from the stack and perform exponentiation.

F_I_DVD

Divides two integers from the stack (and returns integer).

F_I_EXP

Take one float and one integer from the stack and perform exponentiation.

F_MUL

Multiplies two floats from the stack.

F_SUB

Subtracts two floats from the stack.

HALT

Halts the VM/thread.

I_ADD

Adds two integers from the stack.

I_DVD

Divides two integers from the stack.

I_EXP

Take two integers from the stack and perform exponentiation.

I_MUL

Multiplies two integers from the stack.

I_SUB

Subtracts two integers from the stack.

JMPFD

Jumps from data.

JMPFS

Jumps from stack.

PDFS

Puts data from stack.

PDTS

Puts data to stack.

PRTAFD

Print anything from data.

PRTAFS

Print anything from stack.

PRTFD

Prints from data.

PRTFS

Prints from stack.