Struct squid_vm::vminternals::stack::VMStack

pub struct VMStack {
    stack_memory: ArrayVec<Immediates, STACK_SIZE>,
    stack_capacity: usize,
    top: usize,
}

Stack implementation.

Fields
stack_memory: ArrayVec<Immediates, STACK_SIZE>

Contains all the values pushed into the stack.

It’s an ArrayVec so all the data is contained in the stack. Being that it has the speed of a real stack.

stack_capacity: usize

The stack capacity.

top: usize

Points to the latest value pushed into the stack.

Used mostly for monitoring the stack current size.

Implementations

impl VMStack

pub fn new() -> VMStack

Instantiates the VMStack object and returns it.

pub fn get_length(&mut self) -> usize

Used for getting the size of the stack.

pub fn pop(&mut self) -> Immediates

Used for popping a value from the stack and returning it.

pub fn check_empty(&mut self) -> bool

Used to check if the stack is empty at the moment.

pub fn push(&mut self, data: Immediates)

Used for pushing values into the stack.